LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-weight-computed-expected.txt


FAIL Property font-weight value 'normal' computes to '400' assert_equals: expected "400" but got "normal"
FAIL Property font-weight value 'bold' computes to '700' assert_equals: expected "700" but got "bold"
PASS Property font-weight value '100' computes to '100' 
PASS Property font-weight value '200' computes to '200' 
PASS Property font-weight value '300' computes to '300' 
FAIL Property font-weight value '400' computes to '400' assert_equals: expected "400" but got "normal"
PASS Property font-weight value '500' computes to '500' 
PASS Property font-weight value '600' computes to '600' 
FAIL Property font-weight value '700' computes to '700' assert_equals: expected "700" but got "bold"
PASS Property font-weight value '800' computes to '800' 
PASS Property font-weight value '900' computes to '900' 
FAIL 100 made bolder computes to 400 assert_equals: expected "400" but got "normal"
FAIL 200 made bolder computes to 400 assert_equals: expected "400" but got "normal"
FAIL 300 made bolder computes to 400 assert_equals: expected "400" but got "normal"
FAIL 400 made bolder computes to 700 assert_equals: expected "700" but got "bold"
FAIL 500 made bolder computes to 700 assert_equals: expected "700" but got "bold"
PASS 600 made bolder computes to 900 
PASS 700 made bolder computes to 900 
PASS 800 made bolder computes to 900 
PASS 900 made bolder computes to 900 
PASS 100 made lighter computes to 100 
PASS 200 made lighter computes to 100 
PASS 300 made lighter computes to 100 
PASS 400 made lighter computes to 100 
PASS 500 made lighter computes to 100 
FAIL 600 made lighter computes to 400 assert_equals: expected "400" but got "normal"
FAIL 700 made lighter computes to 400 assert_equals: expected "400" but got "normal"
FAIL 800 made lighter computes to 700 assert_equals: expected "700" but got "bold"
FAIL 900 made lighter computes to 700 assert_equals: expected "700" but got "bold"
